mcp[patch]: raise a runtime error if the tool call is not available (#318)
mcp-sdk context manager is swallowing some classes of exceptions with no
way to recover the actual error the occurred.

This patch makes a slight improvement to the exception an end user sees
(so it's not an unbound variable).

Unfortunately, we don't have any information about the root cause (we
don't even know whether a timeout can cause this.)

I'll try to recreate this issue:
https://github.com/langchain-ai/langchain-mcp-adapters/issues/314 in
unit tests to see if we can trigger it with timeouts.