feat: support basic elicitation callback (#385)
It would not be difficult to add support for the other handlers as well.
But we're not seeing much demand yet so I'm not pressed.

Adds support for an `elicitation_handler` to the `MultiServerMCPClient`.
This wraps the `elicitation_handler` supported in the underlying MCP
sdk. We also expose additional context like `server_name` and
`tool_name`.

At the moment, we are not enforcing adherence to `responseSchema`
specified in the elicitation request.

Unfortunately, it's not easy to map elicitation requests onto interrupts
when operating in a LangGraph context. Specifically, the elicitation
handler relies on a blocking call that presumes a stateful, long lived
connection. Interrupts work primarily w/ stateful, short lived
connections.